1st Week Assignments  Outputs (Microservices)
1. What are Primitive Data types in Java ?
Ans ) The eight primitives defined in Java are short, long, float, double, Boolean and char.

2. What are the Identifiers in Java
Ans ) All Java variables must be identified with unique names. These unique names are called identifiers. Like : : ( Ans , Marks , Sum , Add Etc. )

3. What is final class in Java ?
Ans ) The final class is a class that is declared with the final keyword. We can restrict class inheritance by making use of the final class. Final classes cannot be extended or inherited. If we try to inherit a final class, then the compiler throws an error during compilation.

4.  What are the two ways to make a class final ?
Ans ) The First Way to Create Final class is to declare  a class with final keyboard, Another way is to declare all of
its constructors as private. 

5. Can we create an instance of final class in another class ?
Ans ) Yes We can create of instance of final class another class but cannot be inherit a final class , and i know than final keyboard are use to Protect our data or information  like ATM Pin , Password etc.

6. What is Volatile keyword used for ?
Ans ) The volatile keyword prevents the compiler from performing optimization on code involving volatile objects, thus ensuring that each volatile variable assignment and read has a corresponding memory access.

7. What is the use of Transient Keyword ?
Ans ) Transient is a variables modifier used in serialization. At the time of serialization, if we don't want to save value of a particular variable in a file, then we use transient keyword.

8. What are the types of casting. ?
Ans ) I know that there are two types of Casting
 (1) Automatically : -  converting a smaller type to a larger type size. like  ( Byte --> Short --> char --> int --> float --> double. )
 (2) Manually        : -  converting a larger  type to a smaller type size. like  ( double  --> float --> int  -> char -> short -> byte. )

9.  What is boxing and unboxing ?
Ans ) Boxing is the process of converting a primitive datatype into an object wrapper datatype, and unboxing is the process of converting a value from an object wrapper type back to the native primitive value.

10. What is the difference between keywords, identifiers and literals in java ?
Ans ) All Java variables must be identified with unique names . Keyword names are restricted, so you can't use them as variable names. Literals like true , false and null are values that can be assigned, but their names are restricted in the same way that keywords are, i.e. you can't have a variable called true or for .