3rd Week Assignments Output (Microservices)
1) What is the difference between JVM, JRE and JDK ?
Ans -->    (JVM) I know that JVM stands for JAVA VIRTUAL MACHINE    <---
JVM is a heart of Java Programming or It can also run those programs which are written in other languages and compiled to Java bytecode. JVMs are available for many hardware and software platforms. JVM, JRE, and JDK are platform dependent because the configuration of each OS is different from each other. However, Java is platform independent.
--->       (JDK) I Know that JDK Stand's for JAVA DEVELOPMENT KIT     <----
JDK is an implementation of any one of the below given Java Platforms released by Oracle Corporation:
Standard Edition Java Platform
Enterprise Edition Java Platform
Micro Edition Java Platform
--->       (JRE) I Know that JRE Stand's for JAVA RUN TIME ENVIRONMENT     <----
JRE is an acronym for Java Runtime Environment. It is also written as Java RTE. The Java Runtime Environment is a set of software tools which are used for developing Java applications. It is used to provide the runtime environment. It is the implementation of JVM. It physically exists. It contains a set of libraries + other files that JVM uses at runtime.

2) How is Java programming language machine and platform independent ?
Ans ---> I Know that Java is Class Based Object Oriented Programming Language. and Also  Java is platform-independent because it uses a virtual machine. The Java programming language and all APIs are compiled into bytecodes. With The help of (JVM) Bytecodes are effectively platform-independent. Whenever we enter a statement “javac pwSkills.java” or compile a code in java, javac compiles the code. And it creates an intermediate code called Byte Code. Bytecode adds this important feature in Java of being independent/ portable as this bytecode is independent of the platforms. This dependency eases the tasks as it can be used in various environments with lesser planning and translation.

3) Explain how Java programs are Executed by the JVM ?
Ans ---> I know that The JVM is responsible for interpreting and executing Java bytecode, which is a compiled version of the Java source code. It also provides features like memory management, security, and platform independence. Compilation converts source code into bytecode, whereas interpretation executes bytecode directly by the JVM.
.The compilation is done once during the development process, while interpretation happens each time the program is run.
.Compilation allows for the optimization of the code, while interpretation provides greater flexibility and dynamic behavior.
.Overall, both compilation and interpretation play important roles in Java programming, and the choice of which to use will depend on the specific needs of the program.

4) What is the use of Var args ?
Ans ---> Varargs help us avoid writing boilerplate code by introducing the new syntax that can handle an arbitrary number of parameters automatically – using an array under the hood. Varargs are straightforward to use. But there’re a few rules we have to keep in mind:
.Each method can only have one varargs parameter
.The varargs argument must be the last parameter

5) What are the types of Variables in Java ?
Ans ---> I know that there are three types of variables in java: local, instance and static. 
---->   Local Variable  <----
A variable declared inside the body of the method is called local variable. You can use this variable only within that method and the other methods in the class aren't even aware that the variable exists. a local variable cannot be defined with "static" keyword.
--->  Instance Variable  <---
A variable declared inside the class but outside the body of the method, is called an instance variable. It is not declared as static. It is called an instance variable because its value is instance-specific and is not shared among instances.
--->  Static variable <----
A variable that is declared as static is called a static variable. It cannot be local. You can create a single copy of the static variable and share it among all the instances of the class. Memory allocation for static variables happens only once when the class is loaded in the memory.

6) How many types of memory areas are allocated by JVM ?
Ans ---> I know that  the memory area of JVM is broadly divided into five different parts which are method area, heap area, Stack, Program counter (PC) registers area and Native method area.
---> Method Area  <---
Class elements like constant pool, class fields, constructor codes, method codes, etc. Method area can be considered as a part of the heap area but stores per-class data only.
---> Heap Memory <---
Heap Memory in java is used by java runtime to allocate memory to objects and class during a java program’s execution. Whenever an object is created in java, it gets stored into heap memory.
---> Stack Memory  <---
As the name signifies, stack memory is based on LIFO (last in, first out) principle. Stack memory is used for static memory allocation, and each executing thread in a java program has its own stack memory.
---> PC Registers <---
The main function of pc registers is to store the address of currently executing the instruction.
---> Native Area  <---
This area is implemented using languages other than java. With the creation of new threads, memory is allocated in this area for each created thread. The size of the native area can be fixed or dynamic.

7) What is Just in time Compiler (JIT) ? 
Ans --->  I Know that In order to improve performance, JIT compilers interact with the  (JVM) at run time and compile suitable bytecode sequences into native machine code. While using a JIT compiler, the hardware is able to execute the native code, as compared to having the JVM interpret the same sequence of bytecode repeatedly and incurring overhead for the translation process. The JIT compiler is able to perform certain simple optimizations while compiling a series of bytecode to native machine language.